tscache: pull CacheRequest into its own file
The change also includes a small amount of code movement in
`timestamp_cache.go`.
